Refactor: gateway <-> service 레이어 구분
