+++ server/socket/src/gateway/room.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { RoomGateway } from './room.gateway';
+import { RoomService } from './room.service';
 
 @Module({
   imports: [],
   controllers: [],
-  providers: [RoomGateway],
+  providers: [RoomGateway, RoomService],
 })
 export class RoomModule {}
